ecomm-warehouse/requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-http==4.6.0
psycopg2-binary==2.9.9
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pyarrow==13.0.0
pytest==7.4.3
pytest-cov==4.1.0
Flask==3.0.0
